chat-with-google-news/requirements.txt
langchain
langchain-openai
streamlit
langchain-community
gnews
newspaper4k